requirements.txt
anyio==4.9.0
APScheduler==3.11.0
certifi==2025.6.15
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
PySocks==1.7.1
python-dotenv==1.1.1
python-telegram-bot==22.2
pytz==2025.2
sniffio==1.3.1
socksio==1.0.0
SQLAlchemy==2.0.41
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
